🔧 Remove commented code and add ingredient list display
This commit removes commented code in the page.js file and adds functionality to display a list of ingredients for each dish in the dishes.js file.
🔧 Remove commented code and add ingredient list display

This commit removes commented code in the page.js file and adds functionality to display a list of ingredients for each dish in the dishes.js file.