🔧 Fix mask property in CustomSvg component
The mask property in the CustomSvg component was updated to use the maskImage property instead. This change ensures compatibility with different browsers by also including the WebkitMaskImage property.

🔄 Update ShoppingCartItemsList component

- Added increaseQuantity and decreaseQuantity functions from useCart hook.
- Updated itemInfo function to handle null or undefined values.
- Removed commented out code for quantity display options.
- Updated onClick event handlers for quantity buttons to call increaseQuantity and decreaseQuantity functions respectively.

🔧 Fix quantity update in CartProvider

Fixed an issue where the quantity was not being properly updated when decreasing or increasing it. The issue occurred when the quantity reached 1 and further decrease was attempted, resulting in a null value. The fix now checks if the quantity is 1 before decreasing it, and returns early if so.

♻️ Refactor getItemsInCart function

Refactored the getItemsInCart function to clear any invalid values (null, undefined, empty string, empty array) from the itemsInCart array retrieved from local storage. This ensures that only valid items are stored in local storage and used in the application.

🔄 Update countItemsInCart function

Updated countItemsInCart function to handle null or undefined values for item quantities. Now, it correctly counts all valid item quantities in the cart.