🔧 Add data prop to DishesList component
The code changes include adding a new prop called "data" to the DishesList component in two different files. This allows passing data from the parent component to the DishesList component. The purpose of this change is to ensure that the DishesList component receives necessary data for rendering dishes based on selected filters and sorting options.