🔧 Add scroll to top functionality in ButtonFilter component
This commit adds a scroll to top functionality in the ButtonFilter component. When the button is clicked, it will toggle the filter modal and then scroll to the top of the page using `window.scrollTo(0, 0)`. This ensures that users are brought back to the top of the page after interacting with the filter feature.