🔧 Add SkeletonContainer component for loading state
This commit adds a new file `SkeletonContainer.js` which contains the `SkeletonContainer` component. This component is used to display a loading state with skeleton placeholders while data is being fetched. It includes various skeleton elements such as buttons and icons.

The `SkeletonContainer` component is imported and used in the `Page` function of the `page.js` file. When the flag variable is true, the loading state is displayed using the `Suspense` component and the `SkeletonContainer`.