🔍 Update data store for fuzzy search
Update the data store when the data is updated to filter dishes using Fuse.js for fuzzy search. If there is a query and there are results, process the search results by adding scores and sorting them. Set the data store with the scored data. If the query is empty, reset the data store to the full list of dishes. Also update the search terms in the store.

Note: This might be redundant if `setQuery` already updates the search terms in the store.